Refactor types out into types.ts file and use /// <reference types="...">
Fixes packaging problems.
